chore: lint config overhaul, dictionary codec work, and codec lint cleanups
- .golangci.yml: enable default:all with a curated disable set and settings
  (gocritic/gosec/unparam exclusions) tuned for this low-level codec library.
- encoding/chunk: extend the dictionary codec with tests and fuzzing.
- encoding/{bitstream,chunk,compress}, pool: lint-driven cleanups; no
  behavior change.
- remove OPTIMIZATION.md (superseded notes).

+++ encoding/chunk/decimal.go
@@ -279,7 +279,7 @@ func nearestDelta(d, origin int64, precisionBits uint8, prevTZ int) (int64, int)
 		d = -d
 	}
 
-	mask := uint64(^uint64(0)) << tz
+	mask := ^uint64(0) << tz
 
 	d = int64(uint64(d) & mask)
 	if minus {
